gas less options

+++ openood/postprocessors/ft_dist_tta_postprocessor.py
@@ -18,7 +18,9 @@ class DistTTAPostprocessor(FTTTAPostprocessor):
         self.iterations = self.args.iterations_per_phase
         self.switch_phase = self.epochs // 4
 
-        self.ft_checkpoint = Path(config.output_dir) / 'checkpoint_ft.pth'
+        self.ft_checkpoint = None
+        if self.args.gasless:
+            self.ft_checkpoint = Path(config.output_dir) / 'checkpoint_ft.pth'
 
         # number of iterations per phase when no self padding
         min_padded_size = self.chunk_size + sum(self.pad_sizes.values()) - self.pad_sizes.get('self', 0)
@@ -71,12 +73,14 @@ class DistTTAPostprocessor(FTTTAPostprocessor):
     def reset(self, *a, **kw):
 
         super().reset(*a, **kw)
+
         self.ft_checkpoint_loaded = False
-        try:
-            os.remove(self.ft_checkpoint)
-            self.recorder.event('load_ft_state', rm=self.ft_checkpoint.name)
-        except FileNotFoundError:
-            self.recorder.event('load_ft_state', rm='{} does not exist'.format(self.ft_checkpoint.name))
+        if self.ft_checkpoint:
+            try:
+                os.remove(self.ft_checkpoint)
+                self.recorder.event('load_ft_state', rm=self.ft_checkpoint.name)
+            except FileNotFoundError:
+                self.recorder.event('load_ft_state', rm='{} does not exist'.format(self.ft_checkpoint.name))
 
     def epoch_sumup(self, *a, **kw):
         return '[{}]'.format(self.phase[:3]) + super().epoch_sumup(*a, **kw)
@@ -126,7 +130,7 @@ class DistTTAPostprocessor(FTTTAPostprocessor):
         if epoch < self.switch_phase:
             self.phase = 'gas'
 
-        if not epoch:
+        if not epoch and self.ft_checkpoint:
             try:
                 net.load_state_dict(torch.load(self.ft_checkpoint))
                 self.ft_checkpoint_loaded = True
@@ -134,7 +138,7 @@ class DistTTAPostprocessor(FTTTAPostprocessor):
             except FileNotFoundError:
                 self.recorder.event('load_ft_state', no='{} does not exist'.format(self.ft_checkpoint.name))
 
-        if epoch == self.epochs:
+        if epoch == self.epochs and self.ft_checkpoint:
             torch.save(net.state_dict(), self.ft_checkpoint)
 
         if self.ft_checkpoint_loaded and self.phase == 'gas':
